+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 
 ## v0.3
 
+### v0.3.1: Replace BioPython with biotite
+- Replace requirement for local foldseek installation in extract-3di
+- Replace all fasta parsing with biotite.sequence.io.fasta
+
 ### v0.3.0: D-SCRIPT with BMPI
 - Introduce `predict_block` to support multi-GPU prediction of protein interactions with semi-on-the-fly embedding loading, resolving #6 and #11
 - Add sparse embedding loading to `predict_block`, in which only the required subset of each block of embeddings is loaded at each time when predicing only some pairs